# schemas/api/openapi.yaml
openapi: 3.0.4
info:
  title: Smart Home MVP API
  version: 1.0.0
  description: >
    REST API для Smart Home (MVP): аутентификация, устройства, команды, телеметрия.
    Устройства подключаются к MQTT брокеру отдельно (MQTT over TLS); этот API управляет онбордингом и чтением данных.
servers:
  - url: https://api.smarthome.local/api/v1
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    IdempotencyKey:
      type: apiKey
      in: header
      name: Idempotency-Key
  schemas:
    Problem:
      type: object
      description: RFC7807-like error
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
      required: [title, status]
...
    DeviceType:
      type: string
      enum: [Sensor, Relay, Switch, Thermostat, Custom]
    DeviceStatus:
      type: string
      enum: [New, Active, Suspended, Offline]
    Device:
      type: object
      properties:
        id: { type: string, format: uuid }
        homeId: { type: string, format: uuid }
        name: { type: string }
        type: { $ref: "#/components/schemas/DeviceType" }
        status: { $ref: "#/components/schemas/DeviceStatus" }
        lastSeen: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
      required: [id, homeId, name, type, status, createdAt]
    CreateDeviceRequest:
      type: object
      properties:
        homeId: { type: string, format: uuid }
        name: { type: string }
        type: { $ref: "#/components/schemas/DeviceType" }
      required: [homeId, name, type]
...
    PageMeta:
      type: object
      properties:
        next: { type: string, nullable: true, description: "cursor/offset token" }
        limit: { type: integer }
      required: [limit]
paths:
...
  /devices:
    get:
      tags: [Devices]
      summary: Список устройств пользователя/дома
      parameters:
        - in: query
          name: homeId
          schema: { type: string, format: uuid }
          required: false
        - in: query
          name: limit
          schema: { type: integer, default: 50, minimum: 1, maximum: 200 }
        - in: query
          name: cursor
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Device" }
                  page:
                    $ref: "#/components/schemas/PageMeta"
    post:
      tags: [Devices]
      summary: Создать устройство (старт онбординга)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateDeviceRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Device" }
        "400":
          description: Bad Request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Problem" }
...